Separated examples into different binaries
