+++ examples/daqiri_bench_raw_rx_multi_q.yaml
@@ -59,8 +59,3 @@ daqiri:
 
 bench_rx:
   interface_name: "rx_port"
-  gpu_direct: true
-  split_boundary: false
-  batch_size: 10240
-  max_packet_size: 1064
-  header_size: 64
