+++ examples/daqiri_bench_raw_rx_reorder_seq_batch.yaml
@@ -19,7 +19,7 @@ daqiri:
       kind: "device"
       affinity: 0
       num_bufs: 4096
-      buf_size: 262144
+      buf_size: 524288
 
     interfaces:
     - name: "rx_port"
@@ -61,8 +61,3 @@ daqiri:
 
 bench_rx:
   interface_name: "rx_port"
-  gpu_direct: true
-  split_boundary: false
-  batch_size: 1024
-  max_packet_size: 2048
-  header_size: 64
