+++ examples/daqiri_bench_raw_sw_loopback.yaml
@@ -47,16 +47,9 @@ daqiri:
 
 bench_rx:
   interface_name: "loopback_ports"
-  gpu_direct: true
-  split_boundary: false
-  batch_size: 10240
-  max_packet_size: 1064
-  header_size: 64
 
 bench_tx:
   interface_name: "loopback_ports"
-  gpu_direct: true
-  split_boundary: false
   batch_size: 10240
   payload_size: 1000
   header_size: 64
