+++ examples/daqiri_bench_raw_sw_loopback_reorder_seq_1024.yaml
@@ -10,18 +10,18 @@ daqiri:
     loopback: "sw"
 
     memory_regions:
-    - name: "Data_TX_GPU"
-      kind: "device"
+    - name: "Data_TX_CPU"
+      kind: "huge"
       affinity: 0
       num_bufs: 16384
       buf_size: 2048
-    - name: "Data_RX_GPU"
-      kind: "device"
+    - name: "Data_RX_CPU"
+      kind: "huge"
       affinity: 0
       num_bufs: 16384
       buf_size: 2048
-    - name: "Reorder_RX_GPU"
-      kind: "device"
+    - name: "Reorder_RX_CPU"
+      kind: "huge"
       affinity: 0
       num_bufs: 128
       # (2048 - 64) * 1024 = 2,031,616 bytes
@@ -38,7 +38,7 @@ daqiri:
           cpu_core: 11
           timeout_us: 1000
           memory_regions:
-            - "Data_TX_GPU"
+            - "Data_TX_CPU"
           offloads:
             - "tx_eth_src"
       rx:
@@ -50,7 +50,7 @@ daqiri:
           timeout_us: 1000
           batch_size: 1024
           memory_regions:
-            - "Data_RX_GPU"
+            - "Data_RX_CPU"
         flows:
         - name: "flow_0"
           # SW loopback currently does not populate per-packet flow metadata, so use flow id 0.
@@ -63,8 +63,8 @@ daqiri:
             udp_dst: 5000
         reorder_configs:
         - name: "rx_reorder_seq_1024"
-          reorder_type: "gpu"
-          memory_region: "Reorder_RX_GPU"
+          reorder_type: "cpu"
+          memory_region: "Reorder_RX_CPU"
           payload_byte_offset: 64
           flow_ids:
             - 0
@@ -77,16 +77,9 @@ daqiri:
 
 bench_rx:
   interface_name: "loopback_ports"
-  gpu_direct: true
-  split_boundary: false
-  batch_size: 1024
-  max_packet_size: 2048
-  header_size: 64
 
 bench_tx:
   interface_name: "loopback_ports"
-  gpu_direct: true
-  split_boundary: false
   batch_size: 1024
   payload_size: 1000
   header_size: 64
@@ -95,6 +88,5 @@ bench_tx:
   ip_dst_addr: 5.6.7.8
   udp_src_port: 5000
   udp_dst_port: 5000
-  add_sequence_number: true
   sequence_number_offset: 0
   sequence_number_start: 0
