+++ examples/daqiri_bench_raw_tx_rx.yaml
@@ -11,12 +11,12 @@ daqiri:
 
     memory_regions:
     - name: "Data_TX_GPU"
-      kind: "huge"
+      kind: "device"
       affinity: 0
       num_bufs: 51200
       buf_size: 8064
     - name: "Data_RX_GPU"
-      kind: "huge"
+      kind: "device"
       affinity: 0
       num_bufs: 51200
       buf_size: 8064
@@ -57,16 +57,9 @@ daqiri:
 
 bench_rx:
   interface_name: "rx_port"
-  gpu_direct: false
-  split_boundary: false
-  batch_size: 10240
-  max_packet_size: 8064
-  header_size: 64
 
 bench_tx:
   interface_name: "tx_port"
-  gpu_direct: false
-  split_boundary: false
   batch_size: 10240
   payload_size: 8000
   header_size: 64
