+++ examples/daqiri_bench_raw_tx_rx_hds.yaml
@@ -70,16 +70,9 @@ daqiri:
 
 bench_rx:
   interface_name: "rx_port"
-  gpu_direct: true
-  split_boundary: true
-  batch_size: 10240
-  max_packet_size: 1064
-  header_size: 64
 
 bench_tx:
   interface_name: "tx_port"
-  gpu_direct: true
-  split_boundary: true
   batch_size: 10240
   payload_size: 1000
   header_size: 64
