+++ examples/daqiri_bench_raw_tx_rx_reorder_seq_1024_cpu.yaml
@@ -80,16 +80,9 @@ daqiri:
 
 bench_rx:
   interface_name: "rx_port"
-  gpu_direct: false
-  split_boundary: false
-  batch_size: 1024
-  max_packet_size: 2048
-  header_size: 64
 
 bench_tx:
   interface_name: "tx_port"
-  gpu_direct: false
-  split_boundary: false
   batch_size: 1024
   payload_size: 1000
   header_size: 64
@@ -99,8 +92,6 @@ bench_tx:
   udp_src_port: 5000
   udp_dst_port: 5000
 
-  # Writes a monotonic network-order uint32 into UDP payload bytes [0:4]
-  # for each transmitted packet.
-  add_sequence_number: true
+  # Writes a monotonic network-order uint32 into UDP payload bytes [0:4].
   sequence_number_offset: 0
   sequence_number_start: 0
