+++ src/managers/dpdk/daqiri_dpdk_mgr.cpp
@@ -30,6 +30,7 @@
 #include <unordered_map>
 #include <unordered_set>
 #include <regex>
+#include <unistd.h>
 #include <arpa/inet.h>
 #include <cstring>
 #include <memory>
@@ -1448,8 +1449,18 @@ void DpdkMgr::setup_accurate_send_scheduling_mask() {
 std::string DpdkMgr::generate_random_string(int len) {
   const char tokens[] = "abcdefghijklmnopqrstuvwxyz";
   std::string tmp;
+  tmp.reserve(static_cast<size_t>(len));
 
-  for (int i = 0; i < len; i++) { tmp += tokens[rand() % (sizeof(tokens) - 1)]; }
+  static std::atomic<uint64_t> counter{0};
+  uint64_t state =
+      static_cast<uint64_t>(std::chrono::steady_clock::now().time_since_epoch().count());
+  state ^= static_cast<uint64_t>(getpid()) << 32U;
+  state ^= counter.fetch_add(1, std::memory_order_relaxed);
+
+  for (int i = 0; i < len; i++) {
+    state = (state * 2862933555777941757ULL) + 3037000493ULL;
+    tmp += tokens[state % (sizeof(tokens) - 1)];
+  }
 
   return tmp;
 }
